requirements.txt
Django>=1.8
django-filter
django-guardian
djangorestframework
PyYAML
markdown
PyMySQL
mysqlclient
django-cors-headers